% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-function-to-observed-timepoints.R
\name{apply_function_to_observed_timepoints}
\alias{apply_function_to_observed_timepoints}
\title{Apply Function to Subset Drug History}
\usage{
apply_function_to_observed_timepoints(
  drug_history,
  fn = expard::risk_model_current_use()
)
}
\arguments{
\item{drug_history}{The drug history. Unobserved time points are denoted by
\code{NA}}

\item{fn}{A function (Default \code{\link{risk_model_current_use}}).}
}
\value{
Vector of the same length as \code{drug_history}. None observed time points
are denoted by \code{NA}
}
\description{
Observations might not be present for certain patients, since they entered
the data set later time = 1 or before the end. This function applies, for
example, a risk model only to the observed time points.
}
