% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-risk-model-to-observed-timepoints-only.R
\name{apply_risk_model_only_to_observed_time}
\alias{apply_risk_model_only_to_observed_time}
\title{Apply Risk Model to Subset Drug History}
\usage{
apply_risk_model_only_to_observed_time(
  drug_history,
  risk_model = expard::risk_model_current_use()
)
}
\arguments{
\item{drug_history}{The drug history. Unobserved time points are denoted by
\code{NA}}

\item{risk_model}{The risk model (Default \code{\link{risk_model_current_use}}).}
}
\value{
Vector of the same length as \code{drug_history}. None observed time points
are denoted by \code{NA}
}
\description{
Observations might not be present for certain patients, since they entered
the data set later time = 1 or before the end. This function applies
the risk model only to the observed time points.
}
