% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-cohort.R
\name{check_cohort}
\alias{check_cohort}
\title{Check Cohort Object}
\usage{
check_cohort(cohort)
}
\arguments{
\item{cohort}{A cohort}
}
\value{
A \code{cohort} object, see \code{\link{generate_cohort}}
}
\description{
\code{check_cohort} checks whether the argument for a cohort
is valid and turns the object into a \code{cohort} object from
the \code{hccdsim} package, which allows us to use
the output function. In case the input is invalid,
the code halts with an appropriate message.
The minimum requirement is that the \code{cohort}
is a list with the items: \code{drug_history}
and \code{adr_history}. Both should be
\emph{binary} matrix with \code{n_patients} rows and
\code{simulation_time} columns.
}
\seealso{
\code{\link{generate_cohort}}
}
