% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-best-model.R
\name{check_correctness_input}
\alias{check_correctness_input}
\title{Function for checking the correctness of the input given
to find_best_model}
\usage{
check_correctness_input(risk_models, params)
}
\arguments{
\item{risk_models}{}

\item{params}{}
}
\description{
Function for checking the correctness of the input given
to find_best_model
}
