% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create2x2table.R
\name{create2x2table}
\alias{create2x2table}
\title{Creating a \eqn{2 \times 2} Table}
\usage{
create2x2table(drug_ADR_pair, method = c("time-point", "drug-era", "patient"))
}
\arguments{
\item{drug_ADR_pair}{A drug-ADR pair; see \code{\link{generate_cohort}}}

\item{method}{Method used to construct the table; either
\code{time-point}, \code{drug-era} and \code{patient}.
See the description for more information (Default:
\code{time-point})}
}
\value{
A \code{cont_table} object; a list with
\item{\code{a},\code{b},\code{c},\code{d}}{The counts in the table}
\item{\code{method}}{Method that was used to construct the table}
\item{\code{n}}{Total count (\code{n = a + b + c + d}). The interpretation
depends on the \code{method} used}
}
\description{
Creates a \eqn{2 \times 2} contingency table for a
drug and ADR. \cr\cr
A table is structured in the following form:
\tabular{lcc}{
\tab ADR \tab not ADR\cr
drug \tab \code{a} \tab \code{c}\cr
not drug \tab \code{b} \tab \code{d}
}
}
\section{Ways to construct the table}{
The counts can be constructed in three different ways. See for
two of them Zorych et al. (2013).
\subsection{By individual time-points (\code{time-point})}{

Each time-point is counted separately. The counts are the
\emph{number of time points} that
\itemize{
\item{\code{a} - the drug was prescribed and the ADR occurred}
\item{\code{b} - the drug was not prescribed but the ADR occurred}
\item{\code{c} - the drug was prescribed but the ADR did not occur}
\item{\code{d} - the drug was not prescribed and the ADR did not occur}
}
Note that in this case the total count \code{n = a + b + c + d} is
the same as the total number of time points observed, i.e., the total
number of patients times the number of time points observed:
\code{n_patients * simulation_time}.
}

\subsection{By individual patients (\code{patient})}{

In this case, individual patients are counted. The counts are the \emph{number
of patients} that
\itemize{
\item{\code{a} - were prescribed the drug and did experience the ADR }
\item{\code{b} - were never prescribed the drug and did experience the ADR}
\item{\code{c} - were prescribed the drug and never experienced the ADR}
\item{\code{d} - were never prescribed the drug and never experienced the ADR}
}
In this case, the total count \code{n = a + b + c +d} is the same as
the number of patients, \code{n_patients}.
}

\subsection{By individual patients (\code{drug-era})}{

In this case we look at \emph{drug-eras}, i.e., periods in which the
patients was prescribed or not prescribed the drug for a longer time.
For example, if the patient was prescribed the drug from time point 3 to
6, then that period is called a drug-era.
The counts are the \emph{number of drug- and non-drug eras} in which
\itemize{
\item{\code{a} - the drug was prescribed and the ADR occurred}
\item{\code{b} - the drug was not prescribed but the ADR occurred}
\item{\code{c} - the drug was prescribed but the ADR did not occur}
\item{\code{d} - the drug was not prescribed and the ADR did not occur}
}
In this case, the total count \code{n} is the total number of drug- and
non-drug eras.
}
}

\examples{
set.seed(1)
cohort <- generate_cohort(n_patients = 200, n_drug_ADR_pairs = 1)
drug_ADR_pair <- cohort[[1]]

# create the 2x2 contingency table per time-point,
# drug-era and patient:
create2x2table(drug_ADR_pair, method = "time-point")
create2x2table(drug_ADR_pair, method = "drug-era")
create2x2table(drug_ADR_pair, method = "patient")
}
\references{
Zorych, I., Madigan, D., Ryan, P., & Bate, A. (2013). Disproportionality methods for
pharmacovigilance in longitudinal observational databases.
Statistical Methods in Medical Research, 22(1), 39–56.
https://doi.org/10.1177/0962280211403602
}
\seealso{
\code{\link{generate_cohort}}
}
