% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine-first-prescription.R
\name{determine_first_prescription}
\alias{determine_first_prescription}
\title{First Prescription}
\usage{
determine_first_prescription(n_patients, simulation_time, min_chance_drug)
}
\arguments{
\item{n_patients}{Number of patients (Default: 1)}

\item{simulation_time}{The total number of time points for this patient}

\item{min_chance_drug}{The probability of the drug being prescribed
at any given time point}
}
\value{
A time point between \code{1} and \code{simulation_time}
}
\description{
\code{determine_first_perscription} returns a time point
for the first prescription of the drug of interest
given the time frame of the patient (\code{simulation_time})
and the chance for it be prescribed is \code{min_chance_drug}.
The patient, in this case, is always prescribed the drug at least
once. The time point follows a \eqn{\Gamma} distribution.
}
\details{
\emph{Note:} we assume here that the probability of
the drug being prescribed does not depend on any previous prescriptions.
}
\examples{
set.seed(1)
determine_first_prescription(n_patients = 4, simulation_time = 10, min_chance_drug = 0.1)
# -> [1] 2 3 5 9 
}
