% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine-frequency-unique-values.R
\name{determine_frequency_unique_values}
\alias{determine_frequency_unique_values}
\title{Compute Frequency Table}
\usage{
determine_frequency_unique_values(mat, adr_history)
}
\arguments{
\item{mat}{Matrix with relevant values (depends on model)}

\item{adr_history}{History of ADRs}
}
\value{
Tibble
}
\description{
Determine how often different values from time_steps_ago
appear in the data set and how often the ADR occurs for
each of them. This is used for speeding up the computation of the
loglikeihood
}
