% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-best-model.R
\name{determine_number_of_fits}
\alias{determine_number_of_fits}
\title{Auxilary function that determines the number of
models that need to be fitted given the input}
\usage{
determine_number_of_fits(risk_models, params)
}
\description{
Auxilary function that determines the number of
models that need to be fitted given the input
}
