% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate}
\usage{
estimate(
  pair,
  risk_model = expard::risk_model_current_use(),
  start = c(-1, 1),
  method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"),
  control = list()
)
}
\arguments{
\item{pair}{The data for a single drug-ADR pair. See the \code{\link{generate_cohort}}
function. A pair is a list of two matrices, see below}

\item{risk_model}{A risk model function
(Default: \code{risk_model_current_use()})}

\item{start}{Starting point for the base \code{\link{optim}}-solver
(Default: \code{c(-1,1)})}

\item{method}{Methods used by the base \code{\link{optim}}-solver}

\item{control}{List used by the base \code{\link{optim}}-solver}
}
\value{
A model fit. A list with the items
\item{\code{n_patients}}{Total number of patients in the dataset}
\item{\code{simulation_time}}{Total number of time points}
\item{\code{loglikelihood}}{The maximal log-likelihood value}
\item{\code{beta0}}{Intercept for the logistic model}
\item{\code{beta}}{Parameter for the logistic model}
\item{\code{prob_no_adr_with_drug}}{Estimated probability
of no ADR occurring}
\item{\code{prob_ADR_with_drug}}{Estimated probability
of ADR occurring}
\item{\code{convergence}}{Convergence value of \code{\link{optim}}.
\code{0} means that the algorithm converged}
}
\description{
\code{estimate} fits a specified risk model \code{risk_model}.
The maximum likelihood is determined numerically.
}
\section{Pair data oject}{

Minimal requirement is that the cohort is a list with
two matrix with the items
\itemize{
\item{\code{drug_history}  A binary matrix of size
\code{n_patients x simulation_time}  describing
the drug prescriptions.}
\item{\code{adr_history}  A binary matrix of size
\code{n_patients x simulation_time} describing
ADR histories}
}
See \code{\link{check_cohort}} and \code{\link{generate_cohort}}
for more details on the \code{cohort} object.
}

\examples{
cohort <- expard::generate_cohort()

fit <- expard::estimate(cohort[[1]], risk_model = expard::risk_model_current_use())
}
\seealso{
\code{\link{generate_cohort}}
}
