% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expard-package.R
\docType{package}
\name{expard-package}
\alias{expard}
\alias{expard-package}
\title{expard: Drug EXPosures and ADRs}
\description{
An R package for fitting complex drug exposure and adverse drug reaction (ADR) relationships
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bips-hb/expard}
  \item Report bugs at \url{https://github.com/bips-hb/expard/issues}
}

}
\keyword{internal}
