% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-best-model.R
\name{find_best_model}
\alias{find_best_model}
\title{Find the Best Model}
\usage{
find_best_model(
  cohort,
  risk_models = c("immediate", "immediately_after", "extended", "long_time_after"),
  params = list(extension = 1:(cohort$simulation_time - 1), duration =
    1:(cohort$simulation_time - 1), time_since = 1:(cohort$simulation_time - 1)),
  start = c(-1, 1),
  method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"),
  control = list(),
  verbose = TRUE
)
}
\arguments{
\item{cohort}{A cohort, see \code{\link{check_cohort}}}

\item{risk_models}{Risk models that will be applied
(Default: \code{c("immediate","immediately_after","extended","long_time_after")})}

\item{params}{A list with the parameters used for the different risk models}

\item{start, method, control}{Arguments for the built-in \code{\link{optim}}-function}

\item{verbose}{Shows progressbar}
}
\value{
A \code{tibble} data frame
}
\description{
Applies all risk models to a given cohort and determines
the AIC, BIC, the test statistic and \eqn{p}-value.
}
\examples{
cohort <- expard::generate_cohort(n_patients = 100, simulation_time = 8)
find_best_model(cohort, verbose = FALSE)
}
