% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-all-models.R, R/fit-model.R
\name{fit_all_models}
\alias{fit_all_models}
\alias{fit_model}
\title{Fit a Model to Cohort Data}
\usage{
fit_all_models(
  cohort,
  models = c("no-association", "current-use", "past-use", "withdrawal", "delayed",
    "decaying", "delayed+decaying", "long-term"),
  zero_patients = 0,
  zero_timepoints = 0,
  method = c("L-BFGS-B", "Nelder-Mead", "BFGS", "CG", "SANN", "Brent"),
  maxiter = 1000,
  parameters = list(),
  mc.cores = 1
)

fit_model(
  cohort,
  model = c("no-association", "current-use", "past-use", "withdrawal", "delayed",
    "decaying", "delayed+decaying", "long-term"),
  zero_patients = 0,
  zero_timepoints = 0,
  method = c("L-BFGS-B", "Nelder-Mead", "BFGS", "CG", "SANN", "Brent"),
  maxiter = 1000,
  parameters = list()
)
}
\arguments{
\item{cohort}{A cohort dataset. See details below.}

\item{models}{Character vector of model names to fit.}

\item{zero_patients, zero_timepoints}{Integer values. Number of
additional unexposed patients/timepoints to add.}

\item{method}{Methods used by the base \code{\link{optim}}-solver}

\item{maxiter}{Maximum number iterations for the
\code{\link{optim}}-solver}

\item{parameters}{List of additional parameters.}

\item{mc.cores}{Number of cores used for parallelization}

\item{model}{Label for a risk model.}
}
\value{
A model fit. A list with the items
\item{\code{n_patients}}{Total number of patients in the dataset}
\item{\code{simulation_time}}{Total number of time points}
\item{\code{loglikelihood}}{The maximal log-likelihood value}
\item{\code{beta0}}{Intercept for the logistic model}
\item{\code{beta}}{Parameter for the logistic model}
\item{\code{prob_no_adr_with_drug}}{Estimated probability
of no ADR occurring}
\item{\code{prob_ADR_with_drug}}{Estimated probability
of ADR occurring}
\item{\code{convergence}}{Convergence value of \code{\link{optim}}.
\code{0} means that the algorithm converged}
}
\description{
\code{fit_model} fits a specified risk model \code{risk_model}
to cohort data.
}
\section{Cohort data object}{

Minimal requirement is that the cohort is a list with
two matrices:
\describe{
\item{\code{drug_history}}{A binary matrix of size
\code{n_patients x simulation_time} describing
the drug prescriptions.}
\item{\code{adr_history}}{A binary matrix of size
\code{n_patients x simulation_time} describing
ADR histories.}
}
See \code{\link{generate_cohort}} for more details on
the \code{cohort} object.
}

\examples{
\donttest{
set.seed(1)
cohort <- generate_cohort(
  n_patients = 100,
  simulation_time = 50,
  n_drug_ADR_pairs = 1
)
pair <- cohort[[1]]

# fit the no-association model
fit_model(pair, model = "no-association")

# fit the current-use model
fit_model(pair, model = "current-use")
}
}
\seealso{
\code{\link{generate_cohort}}
}
