% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-adr-history.R
\name{generate_adr_history}
\alias{generate_adr_history}
\title{Returns ADR History}
\usage{
generate_adr_history(drug_history, risk_model, min_chance, max_chance, ...)
}
\arguments{
\item{drug_history}{Binary vector denoting the
drug prescriptions history}

\item{risk_model}{A risk model, e.g., \code{\link{risk_model_current_use}}}

\item{min_chance}{The probability of the ADR when
the drug history has no effect}

\item{max_chance}{The probability of the ADR when
the drug history has the highest possible effect}
}
\value{
Binary vector of the same length as \code{drug_history}
}
\description{
Generates a binary vector of ADR occurrences given a drug history
and a risk model.
}
