% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-drug-ADR-pair.R
\name{generate_drug_ADR_pair}
\alias{generate_drug_ADR_pair}
\title{Generate a Drug-ADR pair}
\usage{
generate_drug_ADR_pair(
  simulation_time = 100,
  risk_model = risk_model_current_use(),
  min_chance_drug = 0.1,
  avg_duration = 5,
  max_chance_drug = NULL,
  guaranteed_exposed = TRUE,
  min_chance = 0.1,
  max_chance = 0.4
)
}
\arguments{
\item{simulation_time}{The total number of time steps}

\item{risk_model}{One of the risk models}

\item{min_chance_drug}{The probability of the drug being prescribed
when the drug history and the ADR history have no effect}

\item{avg_duration}{Average number of time points a patient is exposed
once exposed to the drug (Determines \code{max_chance})}

\item{max_chance_drug}{The probability of the ADR when
the drug history and the ADR history have the highest
possible effect (Default: \code{NULL})}

\item{guaranteed_exposed}{If \code{TRUE}, the patient is exposed to the drug
at least once}

\item{min_chance}{The probability of the ADR when
the drug history has no effect}

\item{max_chance}{The probability of the ADR when
the drug history has the highest possible effect}
}
\value{
A list with
\item{\code{drug_history}}{binary vector of the
drug history}
\item{\code{adr_history}}{binary vector of the
ADR history}
}
\description{
Generates the drug- and ADR history
}
\examples{
drug_ADR_pair <- expard::generate_drug_ADR_pair()
}
\seealso{
\code{\link{generate_cohort}}
}
