% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-drug-history.R
\name{generate_drug_history}
\alias{generate_drug_history}
\title{Generate Drug Prescription History}
\usage{
generate_drug_history(
  simulation_time = 10,
  min_chance,
  avg_duration = 5,
  max_chance = NULL,
  guaranteed_exposed = TRUE,
  ...
)
}
\arguments{
\item{simulation_time}{Total number of time points}

\item{min_chance}{The probability of the drug being prescribed
when the drug history and the ADR history have no effect}

\item{avg_duration}{Average number of time points a patient is exposed
once exposed to the drug (Determines \code{max_chance})}

\item{max_chance}{The probability of the ADR when
the drug history and the ADR history have the highest
possible effect (Default: \code{NULL})}

\item{guaranteed_exposed}{If \code{TRUE}, the patient is exposed to the drug
at least once}
}
\value{
\code{1} or \code{0}
}
\description{
Returns whether or not the drug is prescribed
at the next time point (\code{1}) or not (\code{0}).
The probability depends on the drug prescriptions,
the ADR history and the \code{drug_model}.
}
\examples{
generate_drug_history(simulation_time = 10, 
                      min_chance = .1,
                      avg_duration = 5
                      guaranteed_exposed = TRUE)
}
\concept{Update functions}
