% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot the Fit}
\usage{
plot_fit(
  fit,
  x_label = "model",
  title = "",
  y_range = NULL,
  past_values = NULL
)
}
\arguments{
\item{fit}{The fit}

\item{past_values}{If not \code{NULL}, a select number of values for past are
used for the plot}

\item{xlab}{The x-label}
}
\value{
ggplot
}
\description{
Plots the BIC for the different risk models that were fitted by the
function \code{\link{fit_all_models}}.
}
