% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-cohort.R
\name{print.cohort}
\alias{print.cohort}
\title{Print function for \code{\link{generate_cohort}}}
\usage{
\method{print}{cohort}(x, ...)
}
\arguments{
\item{x}{A \code{cohort} object.}

\item{...}{Unused.}
}
\description{
Print function for \code{\link{generate_cohort}}
}
