% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create2x2table.R
\name{print.cont_table}
\alias{print.cont_table}
\title{Print function for 2x2 tables}
\usage{
\method{print}{cont_table}(x, ...)
}
\arguments{
\item{x}{A \code{cont_table} object.}

\item{...}{Unused.}
}
\description{
Print function for 2x2 tables
}
