% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-drug-ADR-pair.R
\name{print.drug_ADR_pair}
\alias{print.drug_ADR_pair}
\title{Function for printing a drug-ADR pair}
\usage{
\method{print}{drug_ADR_pair}(x, ...)
}
\arguments{
\item{x}{A \code{drug_ADR_pair} object.}

\item{...}{Unused.}
}
\description{
Function for printing a drug-ADR pair
}
