% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R, R/fit-model2.R
\name{print.expardfit}
\alias{print.expardfit}
\title{Print function for the expardfit}
\usage{
\method{print}{expardfit}(fit)

\method{print}{expardfit}(fit)
}
\description{
Print function for the expardfit

Print function for the hccd fit_model
}
