% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{print.expardfit2}
\alias{print.expardfit2}
\title{Print function for the hccd fit_model}
\usage{
\method{print}{expardfit2}(x, ...)
}
\arguments{
\item{x}{A \code{expardfit2} object.}

\item{...}{Unused.}
}
\description{
Print function for the hccd fit_model
}
