% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-drug-ADR-pair.R, R/generate-patient.R
\name{print.patient}
\alias{print.patient}
\title{Function for printing a patient}
\usage{
\method{print}{patient}(patient)

\method{print}{patient}(patient)
}
\description{
Function for printing a patient

Function for printing a patient
}
