% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-best-model.R
\name{process_fit_results}
\alias{process_fit_results}
\title{Auxilary function to process a model fit and return
a tibble}
\usage{
process_fit_results(fit, method, n_params)
}
\description{
Auxilary function to process a model fit and return
a tibble
}
