% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-models.R
\name{risk_model_delayed}
\alias{risk_model_delayed}
\title{Risk Model 'Delayed'}
\usage{
risk_model_delayed(mu, sigma)
}
\arguments{
\item{rate}{The rate with which the risk dissipates}
}
\value{
A risk model
}
\description{
Risk Model 'Delayed'
}
\examples{
drug_history <- c(1, 0, 1, 0, 0)

risk_model <- risk_model_withdrawal(rate = 1.2) 
risk_model(drug_history) 
}
\seealso{
Other Risk models: 
\code{\link{risk_model_current_use}()},
\code{\link{risk_model_decaying}()},
\code{\link{risk_model_duration}()},
\code{\link{risk_model_long_term}()},
\code{\link{risk_model_no_association}()},
\code{\link{risk_model_past}()},
\code{\link{risk_model_withdrawal}()}
}
\concept{Risk models}
