% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-models.R
\name{risk_model_long_term}
\alias{risk_model_long_term}
\title{Risk Model 'Long Time After'}
\usage{
risk_model_long_term(rate, delay)
}
\arguments{
\item{rate}{The rate with which the risk increases}

\item{delay}{How long it takes before the risk is .5}
}
\value{
A risk model
}
\description{
A risk model reflects how the probability of
suffering the ADR changes with the drug exposures
of a patient. It returns \code{0} when the drug
prescription history has no effect, and \code{1}
when the patient is at maximal risk.
In this case, the risk of the ADR increases
only after a certain moment in time before it
reaches the maximal risk.
Let \eqn{t_0} be the time point that the drug
was prescribed for the first time, and \eqn{\delta}
be the number of time points since first description
that the function returns 0.5 ("half way"). The
return value is
\deqn{1 / (1 + \exp(-\gamma \cdot (t_0 - \delta)))}
where \eqn{\gamma} is \code{rate}. Note that this
is a sigmoid function.
}
\examples{
drug_history <- c(1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1)

risk_model <- risk_model_long_time_after(rate = .5, delay = 9) 
risk_model(drug_history) 
}
\seealso{
Other Risk models: 
\code{\link{risk_model_current_use}()},
\code{\link{risk_model_decaying}()},
\code{\link{risk_model_delayed}()},
\code{\link{risk_model_duration}()},
\code{\link{risk_model_no_association}()},
\code{\link{risk_model_past}()},
\code{\link{risk_model_withdrawal}()}
}
\concept{Risk models}
