% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-models.R
\name{risk_model_past}
\alias{risk_model_past}
\title{Risk Model 'Withdrawal'}
\usage{
risk_model_past(past)
}
\arguments{
\item{rate}{The rate with which the risk dissipates}
}
\value{
A risk model
}
\description{
A risk model reflects how the probability of
suffering the ADR changes with the drug exposures
of a patient. It returns \code{0} when the drug
prescription history has no effect, and \code{1}
when the patient is at maximal risk.
In this case, once the patient is no longer exposed
to the drug, the probability of the ADR peaks
and then dissipates exponentially.
Let \eqn{\tau} be the number of time points since
the patient was prescribed the drug last. The return
value is the given by
\deqn{\exp(-\gamma \cdot (\tau - 1))}
where \eqn{\gamma} is \code{rate}.
}
\examples{
drug_history <- c(1, 0, 1, 0, 0)

risk_model <- risk_model_withdrawal(rate = 1.2) 
risk_model(drug_history) 
}
\seealso{
Other Risk models: 
\code{\link{risk_model_current_use}()},
\code{\link{risk_model_decaying}()},
\code{\link{risk_model_delayed}()},
\code{\link{risk_model_duration}()},
\code{\link{risk_model_long_term}()},
\code{\link{risk_model_no_association}()},
\code{\link{risk_model_withdrawal}()}
}
\concept{Risk models}
