% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-models.R
\name{risk_model_no_association}
\alias{risk_model_no_association}
\alias{risk_model_current_use}
\alias{risk_model_past}
\alias{risk_model_duration}
\alias{risk_model_withdrawal}
\alias{risk_model_delayed}
\alias{risk_model_decaying}
\alias{risk_model_delayed_decaying}
\alias{risk_model_long_term}
\title{Risk Model 'No Association'}
\usage{
risk_model_no_association()

risk_model_current_use()

risk_model_past(past)

risk_model_duration(duration)

risk_model_withdrawal(rate)

risk_model_delayed(mu, sigma)

risk_model_decaying(rate)

risk_model_delayed_decaying(mu, sigma, rate)

risk_model_long_term(rate, delay)
}
\arguments{
\item{past}{Reference time point > 0.}

\item{duration}{The duration to check against.}

\item{rate}{The rate with which the risk increases}

\item{mu, sigma}{The parameters to \code{\link[=dnorm]{dnorm()}} used to model the risk.}

\item{delay}{How long it takes before the risk is .5}
}
\value{
A risk model
}
\description{
A risk model reflects how the probability of
suffering the ADR changes with the drug exposures
of a patient. It returns \code{0} when the drug
prescription history has no effect, and \code{1}
when the patient is at maximal risk.
In this case, the drug has no effect on the ADR
what so ever.

A risk model reflects how the probability of
suffering the ADR changes with the drug exposures
of a patient. It returns \code{0} when the drug
prescription history has no effect, and \code{1}
when the patient is at maximal risk.
In this case, the probability of the ADR
is maximal only when the patient is currently
exposed.

A risk model reflects how the probability of
suffering the ADR changes with the drug exposures
of a patient. It returns \code{0} when the drug
prescription history has no effect, and \code{1}
when the patient is at maximal risk.
In this case, once the patient is no longer exposed
to the drug, the probability of the ADR peaks
and then dissipates exponentially.
Let \eqn{\tau} be the number of time points since
the patient was prescribed the drug last. The return
value is the given by
\deqn{\exp(-\gamma \cdot (\tau - 1))}
where \eqn{\gamma} is \code{rate}.

A risk model reflects how the probability of
suffering the ADR changes with the drug exposures
of a patient. It returns \code{0} when the drug
prescription history has no effect, and \code{1}
when the patient is at maximal risk.
In this case, once the patient is no longer exposed
to the drug, the probability of the ADR peaks
and then dissipates exponentially.
Let \eqn{\tau} be the number of time points since
the patient was prescribed the drug last. The return
value is the given by
\deqn{\exp(-\gamma \cdot (\tau - 1))}
where \eqn{\gamma} is \code{rate}.

A risk model reflects how the probability of
suffering the ADR changes with the drug exposures
of a patient. It returns \code{0} when the drug
prescription history has no effect, and \code{1}
when the patient is at maximal risk.
In this case, once the patient is no longer exposed
to the drug, the probability of the ADR peaks
and then dissipates exponentially.
Let \eqn{\tau} be the number of time points since
the patient was prescribed the drug last. The return
value is the given by
\deqn{\exp(-\gamma \cdot (\tau - 1))}
where \eqn{\gamma} is \code{rate}.

A risk model reflects how the probability of
suffering the ADR changes with the drug exposures
of a patient. It returns \code{0} when the drug
prescription history has no effect, and \code{1}
when the patient is at maximal risk.
In this case, the risk of the ADR increases
only after a certain moment in time before it
reaches the maximal risk.
Let \eqn{t_0} be the time point that the drug
was prescribed for the first time, and \eqn{\delta}
be the number of time points since first description
that the function returns 0.5 ("half way"). The
return value is
\deqn{1 / (1 + \exp(-\gamma \cdot (t_0 - \delta)))}
where \eqn{\gamma} is \code{rate}. Note that this
is a sigmoid function.
}
\examples{
drug_history <- c(1, 0, 1, 0, 0)
risk_model <- risk_model_no_association()
risk_model(drug_history)
drug_history <- c(1, 0, 1, 0, 0)
risk_model <- risk_model_current_use()
risk_model(drug_history)
drug_history <- c(1, 0, 1, 0, 0)

risk_model <- risk_model_past(past = 2)
risk_model(drug_history)
drug_history <- c(1, 0, 1, 0, 0)

risk_model <- risk_model_duration(duration = 2)
risk_model(drug_history)
drug_history <- c(1, 0, 1, 0, 0)

risk_model <- risk_model_withdrawal(rate = 1.2)
risk_model(drug_history)
drug_history <- c(1, 0, 1, 0, 0)

risk_model <- risk_model_delayed(5, 1)
risk_model(drug_history)
drug_history <- c(1, 0, 1, 0, 0)

risk_model <- risk_model_decaying(rate = 1.2)
risk_model(drug_history)
drug_history <- c(1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1)

risk_model <- risk_model_long_term(rate = .5, delay = 9)
risk_model(drug_history)
}
\concept{Risk models}
