% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create2x2tables.R
\name{create2x2tables}
\alias{create2x2tables}
\title{Creating \eqn{2 \times 2} Tables}
\usage{
create2x2tables(
  cohort,
  method = c("time-point", "drug-era", "patient"),
  verbose = TRUE
)
}
\description{
Creating \eqn{2 \times 2} Tables
}
