% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-cohort.R
\name{generate_cohort}
\alias{generate_cohort}
\title{Generate a Cohort}
\usage{
generate_cohort(
  n_patients = 100,
  simulation_time = 100,
  n_drug_ADR_pairs = 50,
  risk_model = rep("risk_model_current_use()", n_drug_ADR_pairs),
  min_chance_drug = rep(0.1, n_drug_ADR_pairs),
  avg_duration = rep(5, n_drug_ADR_pairs),
  max_chance_drug = rep(NULL, n_drug_ADR_pairs),
  prob_guaranteed_exposed = rep(1, n_drug_ADR_pairs),
  min_chance = rep(0.1, n_drug_ADR_pairs),
  max_chance = rep(0.4, n_drug_ADR_pairs),
  verbose = FALSE
)
}
\arguments{
\item{n_patients}{Number of patients}

\item{simulation_time}{The total number of time steps}

\item{risk_model}{One of the risk models}

\item{min_chance_drug}{The probability of the drug being prescribed
when the drug history and the ADR history have no effect}

\item{max_chance_drug}{The probability of the ADR when
the drug history and the ADR history have the highest
possible effect}

\item{min_chance}{The probability of the ADR when
the drug history has no effect}

\item{max_chance}{The probability of the ADR when
the drug history has the highest possible effect}

\item{verbose}{Show progress bar (Default: \code{FALSE})}
}
\value{
A \code{cohort} object; a list with
\item{\code{n_patients}}{The total number of patients}
\item{\code{simulation_time}}{The total number of time steps}
\item{\code{drug_history}}{A binary matrix with \code{n_patients}
rows and \code{simulation_time} columns}
\item{\code{adr_history}}{A binary matrix with \code{n_patients}
rows and \code{simulation_time} columns}
\item{\code{patient_profile}}{A list with the patient profiles for each patient}
}
\description{
Generates a cohort of patients, each with a
drug prescription and ADR history.
See for more information and details \code{\link{generate_patient}}.
}
\examples{
# choose a patient model (in this case based on the patients sex)
patient_model <- patient_model_sex(prob_male = 0.5)
(patient_profile <- patient_model())

# choose a risk model (effect of the drug is immediate)
risk_model <- risk_model_current_use()

# how the drug is prescribed over time
drug_model <- drug_model_markov_chain() 

# how the occurence of an ADR affect the prescription of the drug
adr_model = adr_model_no_effect()

# minimum and maximal probabilities for the drug being prescribed
min_chance_drug <- probability_model_sex(prob_male = .1, prob_female = .2)
max_chance_drug <- probability_model_sex(prob_male = .5, prob_female = .53)

# minimum and maximal probabilities for the ADR occurring 
min_chance_adr <- probability_model_sex(prob_male = .01, prob_female = .05)
max_chance_adr <- probability_model_sex(prob_male = .2, prob_female = .3)

# chance that the patient is exposed to the drug
prob_exposure <- probability_model_sex(prob_male = .9, prob_female = .7)

generate_cohort(n_patients = 10, 
               simulation_time = 20, 
               risk_model      = risk_model, 
               drug_model      = drug_model,
               adr_model       = adr_model, 
               min_chance_drug = min_chance_drug,
               max_chance_drug = max_chance_drug,
               min_chance_adr  = min_chance_adr, 
               max_chance_adr  = max_chance_adr, 
               patient_model   = patient_model,
               verbose = TRUE) 
}
\seealso{
\code{\link{generate_patient}}
}
