% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-drug-history.R
\name{generate_drug_history}
\alias{generate_drug_history}
\title{Generate Drug Prescription History}
\usage{
generate_drug_history(simulation_time = 10, min_chance, max_chance, ...)
}
\arguments{
\item{simulation_time}{Total number of time points}

\item{min_chance}{The probability of the drug being prescribed
when the drug history and the ADR history have no effect}

\item{max_chance}{The probability of the ADR when
the drug history and the ADR history have the highest
possible effect}

\item{drug_model}{A drug model function,
e.g., \code{\link{drug_model_markov_chain}}}
}
\value{
\code{1} or \code{0}
}
\description{
Returns whether or not the drug is prescribed
at the next time point (\code{1}) or not (\code{0}).
The probability depends on the drug prescriptions,
the ADR history and the \code{drug_model}.
}
\examples{
# choose the drug prescription model
drug_model <- drug_model_markov_chain()

# choose probability models. Note that the probabilities change
# with sex
min_chance <- probability_model_sex(prob_male = .1, prob_female = .05)
max_chance <- probability_model_sex(prob_male = .7, prob_female = .8)

# create a patient profile: 
patient_model <- patient_model_sex(prob_male = 0.5) 
(patient_profile <- patient_model()) 

generate_drug_history(simulation_time = 10, 
                      drug_model, 
                      min_chance,
                      max_chance,
                      patient_profile)
}
\seealso{
Other Update functions: 
\code{\link{generate_adr_history}()}
}
\concept{Update functions}
