% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Negative of the Log-Likelihood Function}
\usage{
loglikelihood(betas, risks, adr_history)
}
\arguments{
\item{betas}{Vector with two entries: \code{beta0, beta}}

\item{risks}{Risks determined by the risk model. The risk
model was applied to the observed drug prescription
history}
}
\description{
\code{loglikelihood} returns the negative
log-likelihood function, defined as
\deqn{\pi = \frac{\exp(\beta_0 + \code{risks}\cdot \beta)}{1 + \exp(\beta_0 + \code{risks}\cdot \beta)}}
where \eqn{\pi} is the probability.
}
