% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot the Fit}
\usage{
plot_fit(fit, xlab = "model", past_range = NULL)
}
\arguments{
\item{fit}{The fit}

\item{xlab}{The x-label}

\item{past}{If not \code{NULL}, a select number of values for past are
used for the plot}
}
\value{
ggplot
}
\description{
Plots the BIC for the different risk models that were fitted by the
function \code{\link{fit_all_models}}.
}
