% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create2x2table.R, R/create2x2tables.R
\name{print.cont_table}
\alias{print.cont_table}
\title{Print function for 2x2 tables}
\usage{
\method{print}{cont_table}(table)

\method{print}{cont_table}(table)
}
\description{
Print function for 2x2 tables

Print function for 2x2 tables
}
