% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-drug-ADR-pair.R
\name{print.drug_ADR_pair}
\alias{print.drug_ADR_pair}
\title{Function for printing a drug-ADR pair}
\usage{
\method{print}{drug_ADR_pair}(drug_ADR_pair)
}
\description{
Function for printing a drug-ADR pair
}
